requirements.txt
#cuda==11.1
numpy==1.19.2
librosa==0.8.0 
joblib==1.0.1
pyyaml==5.4.1
tensorboardX==2.2
SoundFile
